refactor(simulation): extract scenario command factory
